refactor(query_api): extract graph + impact query surface to graph.rs
Move ffi_surface, find_callers/trace_callees (+ _with_options + traverse/enrich/
graph_logical_symbol/graph_options_with_logical_group helpers), graph_traversal_report,
compare_graph_to_text, compare_graph_to_scip, local_symbol_context_hits, and the
impact_surface* methods into graph.rs (the graph-backed query job family). The two
oracle-label free fns stay in mod.rs; graph.rs imports them via super::. Helpers
called from outside (local_symbol_context_hits, graph_options_with_logical_group)
are pub(super). No behavior change.
